Direct user to example+docs after Heroku deploy
It would be great to show users how the API works upon successful deploy to Heroku.  Setting the `success_url` parameter in `app.json` will open up the specified URL on successful deploy.  To make it even more useful, I set it to the README for this project so the user can continue understanding how the API works.