Adding Dockerfile
This adds docker support to this app. It uses a modern Node image and only copies into the container the files necessary for building the project. It then mounts the source files at run time. This allows images to be reused between most commits without having to rebuild the whole Docker image.