Update (#11)
* Direct user to example+docs after Heroku deploy

It would be great to show users how the API works upon successful deploy to Heroku.  Setting the `success_url` parameter in `app.json` will open up the specified URL on successful deploy.  To make it even more useful, I set it to the README for this project so the user can continue understanding how the API works.

* Add missing comma

* Try to fix markdown formatting

* Allow for an attachment name to be specified in order to receive rendering as a downloadable attachment

* Style fixes

* change content length check, fixes failing Google unit test

* feat: optionally fail if there are failed requests

There are use-cases where the pdf is requested and processed
automatically, e.g. creation of personalised pdf which are sent out in
newsletters.

If the page or parts of it fail to load, then it may be preferable to
not create a PDF and report an error instead of creating a pdf of an
error page or of a page with, for example, missing images or missing
fonts.

This commit introduces a new option 'failEarly'. This can take the
values 'all' or 'page'. If any other value is used or the option is
omitted then the previous behavior is unchanged.
If the value for 'failEarly' is 'all', then the process fails with an 412
http error code and informs the client how many requests have failed.
Requests are considered failed if they returned a response with an http
status code equal or larger than 400.
If the value for 'failEarly' is 'page', then the process fails when the
request for the provided url results in any response other than 200.
This explicitly includes responses with the exit code of 302, which
might otherwise produce a valid PDF.

* allow to set charset in header of post request to api

* Add the most minimal example to readme

* Test URI changed to github.com to avoid google banhammer in Jenkins

* Using keep-alive in tests

* update puppeteer version 1.6.0

* resolve travis ci errors

* Fix support for footer and header templates in GET request

* Document footer/header templates

* Update warnings for footer/header templates

* Add charset utf-8 when rendering HTML to fix rendering emojis

* Add more info to security warning

* Add warning emojis

* Add test cases for larger images

* Implement screenshot rendering support

* Document screenshot options

* Remove unnecessary stripe status checks

* Merging

* fix

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug