Upgrade puppeteer to 1.14
The motivation behind the switch from `pdf2json` to `pdf-parse` was that the
former was not able to correctly read the PDF:s generated by the updated
Chromium version. Looking at the GitHub issues of that project showed that it
was quite buggy and not really maintained anymore. `pdf-parse` on the other
hand had no problem with PDF parsing, so the choice was easy.